Clean up project structure
